+++ CloudShell.Hosting/ResourceManager/ResourceManagerUiOptions.cs
@@ -5,4 +5,8 @@ public sealed class ResourceManagerUiOptions
     public const string SectionName = "ResourceManager";
 
     public bool ReadOnly { get; set; }
+
+    public bool ShowRuntimeManagedResources { get; set; }
+
+    public bool ShowHiddenResources { get; set; }
 }